Add `enabled` property on Logger
Adds a property on `Logger` to enabled logging. When ever `log(error:
Error)` is called it will check if the feature is enabled or not. If
enabled it will print to the console. This will no use `NSLog` instead
of `print` as that will log to the device console which `print` does
not. It should improve the debugging experience if you debug on a
device using the device console.