Implement expiration mode on `Config`
Implements a new configuration option for opting out from using the
internal expiration cleanup mechanism in `Cache`.

Expiration mode lives on `Config` and is used by `CacheManager`. When
the expiration mode is set to `.auto`, `Cache` will handle the cleanup
of expired objects. When set to `.manual`, it means that you will
manually handle expired objects.